Fix handle alignment problem in Safari.
Actually, the safari problem alerted me to an ever deeper problem that
was happening everywhere but was only showing up in Safari.  Since the
audio player itself manipulates the handle via the "left" CSS
attribute, it makes sense to align the gutter so that the beginning
position of the handle is at 0, not some negative offset that doesn't
need to actually be there.

Ensure handle always starts in the same place.