Defer loading the createPlayer function.
Though I don't totally know what is happening with "Defer" (and
apparently the syntax is not entirely clear also: is it just "defer"
or "defer='defer'"?) but it seems like sometimes the page is trying to
call createPlayer before master.js is loaded so there's an error.
Seems like this might fix it.